+++ src/Stopwatch.h
@@ -1,3 +1,13 @@
+/*
+* A Stopwatch is used for time measurement. You can start it, and get elapsed
+* time in ms since start.
+* As a convenience funtion, getNextFrameDelay(int frametime) returns a time
+* difference between frametime an elapsed time, floored by 10ms.
+*
+*  Created on: Nov 10, 2016
+*      Author: Edwin Huber
+*/
+
 #pragma once
 
 #include <opencv2/core/core.hpp>
@@ -11,8 +21,8 @@ private:
 public:
 	Stopwatch();
 	~Stopwatch();
-	int getElapsedTime();
-	int getNextFrameDelay(int frametime);
+	int64 getElapsedTime();
+	int64 getNextFrameDelay(int frametime);
 	void reset();
 };
